% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emis_grid_array.R
\name{emis_grid_array}
\alias{emis_grid_array}
\title{Similar to vein::emis_grid but it returns an array with dimensions of ppchem::read_edgar}
\usage{
emis_grid_array(spobj, month, sf = FALSE, sr = 4326, type = "lines")
}
\arguments{
\item{spobj}{A spatial dataframe of class "sp" or "sf". When class is "sp"
it is transformed to "sf". It can be points or lines.}

\item{month}{Integer, which month? from 1 to 12.}

\item{sf}{Logical Do you want an 'sf' instead an array?}

\item{sr}{Integer, Spatial reference e.g: 31983. It is required if spobj and g are
not projected. Please, see http://spatialreference.org/.}

\item{type}{type of geometry: "lines" or "points".}
}
\description{
\code{\link{emis_grid_array}} return array or sf object of gridded emisions.
}
\note{
When spobj is a 'Spatial' object (class of sp), they are converted
 into 'sf'. Also, The aggregation of data ise done with data.table functions.
}
\examples{
\dontrun{
# do not run
library(vein)
data(net)
net$emission <- 1:length(net)
netg <- emis_grid_array(spobj = net[,"emission"])
}
}
