% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_edgar.R
\name{read_edgar}
\alias{read_edgar}
\title{Read HDF5 EDGAR emission files.}
\usage{
read_edgar(path, dataset, show.attribute = FALSE, month,
  verbose = TRUE)
}
\arguments{
\item{path}{Character; Where is your HDF5 file.}

\item{dataset}{Character; What data do you want. If you dont know,
i will show you that it is inside so that you can choose.}

\item{show.attribute}{Logical; Do you wna to see attirbutes?. If you dont know which,
i will show you that it is inside so that you can choose.}

\item{month}{Integer; which month from 1 to 12.}

\item{verbose}{Logical; Do you want more information?}
}
\value{
array of EDGAR emissions by month.
}
\description{
\code{\link{read_edgar}} calculates ADT based on hourly traffic data. The
input traffic data is usually for morning rush hours.
}
\examples{
\dontrun{
# do not run
path  = "~/Downloads/EDGAR-HTAP_TRANSPORT_2010.h5"
e <- read_edgar(path)
}
}
