% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_edgar.R
\name{write_edgar}
\alias{write_edgar}
\title{Read HDF5 EDGAR emission files.}
\usage{
write_edgar(path, dataset, m, month, verbose = TRUE)
}
\arguments{
\item{path}{Character; Where is your HDF5 file.}

\item{dataset}{Character; What data do you want. If you dont know,
i will show you that it is inside so that you can choose.}

\item{m}{Array; Array of monhly emissions, dim = c(month, 1, , )}

\item{month}{Integer; which month from 1 to 12. If missing, cover whole year.}

\item{verbose}{Logical; Do you want more information?}
}
\description{
\code{\link{read_edgar}} calculates ADT based on hourly traffic data. The
input traffic data is usually for morning rush hours.
}
\examples{
\dontrun{
# do not run
path  = "~/Downloads/EDGAR-HTAP_TRANSPORT_2010.h5"
e <- read_edgar(path, month = 1, dataset = "PM2.5")
e <- e* #units in kg/m2/s
write_edgar(path = path, dataset = "PM2.5", month = 1)
}
}
